+++ tests/stubs/Queries/ThrowError.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Butler\Graphql\Tests\Queries;
+
+use Error;
+
+class ThrowError
+{
+    public function __invoke($root, $args, $context)
+    {
+        throw new Error('error');
+    }
+}